Allow partials to define blocks and be compiled
I have an instance where I want to be able to compile partials individually. The partials declare their own blocks so that they can be extended by other partials. Currently when trying to compile partials in this way an error occurs because the context has no _blocks, the little tweak above fixes this issue.